requirements.txt
numpy==1.19.3
pandas==1.1.3
python-decouple==3.6
PyYAML==6.0
requests==2.24.0
SQLAlchemy==1.3.20
SQLAlchemy_Utils==0.38.2